manager: left-align WebUI button and add dedicated webui icon (#1296)
The WebUI button is now positioned on the left beside the module information.
Only the WebUI button layout changed; the update and uninstall/action-script buttons remain right-aligned.
This is done by moving the WebUI button to the section before the weighted Spacer.
The old settings icon previously used for WebUI has been replaced with a new dedicated icon.
A new `webui.xml` vector drawable is added to provide this updated WebUI icon.

Signed-off-by: Prslc <prslc113@gmail.com>